requirements.txt
fastapi==0.114.2
opencv-python==4.10.0.84
ultralytics==8.2.95
numpy==1.26.4
torch==2.4.1
pydantic==2.9.2
pillow==10.4.0
uvicorn==0.31.1